+++ data/crac/crac-impl/src/main/java/com/powsybl/openrao/data/crac/impl/InjectionRangeActionAdderImpl.java
@@ -7,8 +7,6 @@
 
 package com.powsybl.openrao.data.crac.impl;
 
-import com.powsybl.iidm.network.Generator;
-import com.powsybl.iidm.network.Load;
 import com.powsybl.iidm.network.Network;
 import com.powsybl.openrao.commons.OpenRaoException;
 import com.powsybl.openrao.data.crac.api.NetworkElement;
@@ -16,7 +14,6 @@ import com.powsybl.openrao.data.crac.api.rangeaction.InjectionRangeAction;
 import com.powsybl.openrao.data.crac.api.rangeaction.InjectionRangeActionAdder;
 
 import java.util.*;
-import java.util.stream.Collectors;
 
 import static com.powsybl.openrao.commons.logs.OpenRaoLoggerProvider.BUSINESS_WARNS;
 import static com.powsybl.openrao.data.crac.impl.AdderUtils.assertAttributeNotEmpty;
@@ -74,7 +71,7 @@ public class InjectionRangeActionAdderImpl extends AbstractStandardRangeActionAd
         runCheckBeforeAdding();
 
         Map<NetworkElement, Double> neAndDk = addNetworkElements();
-        this.initialSetpoint = getCurrentSetpoint(network, neAndDk);
+        this.initialSetpoint = InjectionRangeActionUtils.getCurrentSetpoint(network, neAndDk);
         InjectionRangeAction injectionRangeAction = new InjectionRangeActionImpl(this.id, this.name, this.operator, this.groupId, this.usageRules, this.ranges, this.initialSetpoint, neAndDk, speed, activationCost, variationCosts);
         this.getCrac().addInjectionRangeAction(injectionRangeAction);
         return injectionRangeAction;
@@ -106,41 +103,6 @@ public class InjectionRangeActionAdderImpl extends AbstractStandardRangeActionAd
         // check usage rules
         if (usageRules.isEmpty()) {
             BUSINESS_WARNS.warn("InjectionRangeAction {} does not contain any usage rule, by default it will never be available", id);
-        }
-    }
-
-    public static double getCurrentSetpoint(Network network, Map<NetworkElement, Double> injectionDistributionKeys) {
-        List<Double> currentSetpoints = injectionDistributionKeys.entrySet().stream()
-            .map(entry -> getInjectionSetpoint(network, entry.getKey().getId(), entry.getValue()))
-            .collect(Collectors.toList());
-
-        if (currentSetpoints.size() == 1) {
-            return currentSetpoints.get(0);
-        } else {
-            Collections.sort(currentSetpoints);
-            if (Math.abs(currentSetpoints.get(0) - currentSetpoints.get(currentSetpoints.size() - 1)) < 1) {
-                return currentSetpoints.get(0);
-            } else {
-                throw new OpenRaoException(String.format("Cannot evaluate reference setpoint of InjectionRangeAction, as the injections are not distributed according to their key"));
-            }
-        }
-    }
-
-    public static double getInjectionSetpoint(Network network, String injectionId, double distributionKey) {
-        Generator generator = network.getGenerator(injectionId);
-        if (generator != null) {
-            return generator.getTargetP() / distributionKey;
-        }
-
-        Load load = network.getLoad(injectionId);
-        if (load != null) {
-            return -load.getP0() / distributionKey;
-        }
-
-        if (network.getIdentifiable(injectionId) == null) {
-            throw new OpenRaoException(String.format("Injection %s not found in network", injectionId));
-        } else {
-            throw new OpenRaoException(String.format("%s refers to an object of the network which is not an handled Injection (not a Load, not a Generator)", injectionId));
         }
     }
