+++ data/crac/crac-impl/src/main/java/com/powsybl/openrao/data/crac/impl/InjectionRangeActionImpl.java
@@ -19,11 +19,9 @@ import com.powsybl.iidm.network.Generator;
 import com.powsybl.iidm.network.Load;
 import com.powsybl.iidm.network.Network;
 
-import java.util.Collections;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
-import java.util.stream.Collectors;
 
 /**
  * @author Baptiste Seguinot {@literal <baptiste.seguinot at rte-france.com>}
@@ -114,38 +112,7 @@ public class InjectionRangeActionImpl extends AbstractRangeAction<InjectionRange
 
     @Override
     public double getCurrentSetpoint(Network network) {
-        List<Double> currentSetpoints = injectionDistributionKeys.entrySet().stream()
-                .map(entry -> getInjectionSetpoint(network, entry.getKey().getId(), entry.getValue()))
-                .collect(Collectors.toList());
-
-        if (currentSetpoints.size() == 1) {
-            return currentSetpoints.get(0);
-        } else {
-            Collections.sort(currentSetpoints);
-            if (Math.abs(currentSetpoints.get(0) - currentSetpoints.get(currentSetpoints.size() - 1)) < 1) {
-                return currentSetpoints.get(0);
-            } else {
-                throw new OpenRaoException(String.format("Cannot evaluate reference setpoint of InjectionRangeAction %s, as the injections are not distributed according to their key", this.getId()));
-            }
-        }
-    }
-
-    public double getInjectionSetpoint(Network network, String injectionId, double distributionKey) {
-        Generator generator = network.getGenerator(injectionId);
-        if (generator != null) {
-            return generator.getTargetP() / distributionKey;
-        }
-
-        Load load = network.getLoad(injectionId);
-        if (load != null) {
-            return -load.getP0() / distributionKey;
-        }
-
-        if (network.getIdentifiable(injectionId) == null) {
-            throw new OpenRaoException(String.format("Injection %s not found in network", injectionId));
-        } else {
-            throw new OpenRaoException(String.format("%s refers to an object of the network which is not an handled Injection (not a Load, not a Generator)", injectionId));
-        }
+        return InjectionRangeActionUtils.getCurrentSetpoint(network, injectionDistributionKeys);
     }
 
     @Override
